Move camping gear dashboard to top-level nav and clean up URL (#2151)
Tessa requested this be more prominent than a Reports sub-item since
librarians use it as an operational daily view. Moved it next to
Appointments in the top-level nav and changed the URL from
/admin/reports/myturn_reservations to /admin/camping_gear. Added a
redirect from the old URL so existing bookmarks still work.
